chore: parse long not int from console input for consistency

+++ src/com/company/Main.java
@@ -4,6 +4,6 @@ public class Main {
 
     public static void main(String[] args){
         BinaryComparator comparator = new BinaryComparator(new SimpleGenerator(16807), new SimpleGenerator(48271));
-        System.out.println(comparator.compareNumbers(Integer.parseInt(args[0]), Integer.parseInt(args[1])));
+        System.out.println(comparator.compareNumbers(Long.parseLong(args[0]), Long.parseLong(args[1])));
     }
 }
